Add a download replay modal for first-time users landing on the library in a non-firefox browser
