+++ src/ui/components/shared/OnboardingModal/DownloadReplayPromptModal.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import DownloadReplayModal from "./DownloadReplayModal";
+
+export default function DownloadReplayPromptModal() {
+  return (
+    <DownloadReplayModal>{`To start making recordings, you'll have to download the Replay browser. Would you like to go that team, or download Replay?`}</DownloadReplayModal>
+  );
+}
